requirements.txt
tensorboard
scikit-learn
seqeval
psutil
sacrebleu
rouge-score
tensorflow_datasets
pytorch-lightning==0.8.5
matplotlib
git-python==1.0.3
streamlit
elasticsearch
transformers
pandas
pytest
conllu
spacy
inflect
wandb
bert-score
tabulate
nltk==3.5
numpy